+++ lyrebird/src/main/java/moe/lyrebird/view/viewmodel/tokenization/extractors/MentionsTokensExtractor.java
@@ -1,4 +1,4 @@
-package moe.lyrebird.view.viewmodel.tokenization.tokenizers;
+package moe.lyrebird.view.viewmodel.tokenization.extractors;
 
 import java.util.Arrays;
 import java.util.List;
